% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_sparse_algo.r
\name{ElasticNetwSVM}
\alias{ElasticNetwSVM}
\title{Function to perform ElasticNet wSVM}
\usage{
ElasticNetwSVM(
  x.train,
  y.train,
  x.tune,
  y.tune,
  x.test,
  y.test,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  l2.option = TRUE,
  beta_precision,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{traing x}

\item{y.train}{traing y}

\item{x.tune}{tuning x}

\item{y.tune}{tuning y}

\item{x.test}{testing x}

\item{y.test}{testing y}

\item{kernel}{Kernel params list. default: 'linear' in sparse learning}

\item{l2.option}{option to use L2-wSVMs for probability estimation}

\item{beta_precision}{round decimal points for beta}

\item{eps}{Small threshold}
}
\value{
Result matrix for downstream analysis
}
\description{
Function to perform ElasticNet wSVM
}
