% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_algo.r
\name{L2wSVM}
\alias{L2wSVM}
\title{Function to perform L2 wSVM}
\usage{
L2wSVM(
  x.train,
  y.train,
  x.tune,
  y.tune,
  x.test,
  y.test,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  small.tune = TRUE,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{traing x}

\item{y.train}{traing y}

\item{x.tune}{tuning x}

\item{y.tune}{tuning y}

\item{x.test}{testing x}

\item{y.test}{testing y}

\item{kernel}{Kernel params list. default: 'linear' in sparse learning}

\item{small.tune}{size of lambda to tune}

\item{eps}{Small threshold}
}
\value{
Result matrix for downstream analysis
}
\description{
Function to perform L2 wSVM
}
