% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{QP.compact.matrix}
\alias{QP.compact.matrix}
\title{find Amat.compact and Aind for QP.solve.compact}
\usage{
QP.compact.matrix(Amat)
}
\arguments{
\item{Amat}{Amat in solve.QP}
}
\value{
compact form of Amat
}
\description{
find Amat.compact and Aind for QP.solve.compact
}
