% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enp_wsvm.r
\name{wENp.svm}
\alias{wENp.svm}
\title{Function to perform Elastic Net weighted SVM using primal optimization}
\usage{
wENp.svm(
  x.train,
  y.train,
  PI = 0.5,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  lambda1 = 0.01,
  lambda2 = 0.01,
  beta_precision = 0.001,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{Training data in matrix form}

\item{y.train}{Label as vector}

\item{PI}{Weight on negative class data}

\item{kernel}{Kernel to choose, refer to function wsvm.kernel, default is linear kernel}

\item{lambda1}{Regulatory parameter for L1-norm penalty and for solving elatic netweighted SVMs, lambda1 > 0}

\item{lambda2}{Regulatory parameter for L2-norm penalty and for solving elatic netweighted SVMs, lambda2 > 0}

\item{beta_precision}{minimum value of beta before set to 0}

\item{eps}{Small threshold}
}
\value{
Elastic Net regulated wSVM model object use primal optimization
}
\description{
Function to perform Elastic Net weighted SVM using primal optimization
}
