% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{wENsvm.predict}
\alias{wENsvm.predict}
\title{predict the label with given elastic net wsvm classifier}
\usage{
wENsvm.predict(
  x,
  beta,
  beta0,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL)
)
}
\arguments{
\item{x}{new data points}

\item{beta}{parameter beta}

\item{beta0}{parameter beta0}

\item{kernel}{Kernel to use, default 'linear'}
}
\value{
return Predicted label {+1,-1}
}
\description{
predict the label with given elastic net wsvm classifier
}
