% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1p_wsvm.r
\name{wl1p.svm}
\alias{wl1p.svm}
\title{Function to perform L1 weighted SVM using primal approach}
\usage{
wl1p.svm(
  x.train,
  y.train,
  PI = 0.5,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  lambda = 0.01,
  beta_precision = 1e-10
)
}
\arguments{
\item{x.train}{Training data in matrix form}

\item{y.train}{Label as vector}

\item{PI}{Weight on negative class data}

\item{kernel}{Kernel to choose, refer to function wsvm.kernel, default is linear kernel}

\item{lambda}{Regulatory parameter for L1-norm penalty and for solving weighted SVMs, lambda > 0}

\item{beta_precision}{minimum value of beta before set to 0}
}
\value{
L1-norm regulated wSVM model object use primal optimization
}
\description{
Function to perform L1 weighted SVM using primal approach
}
