% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_wsvm.r
\name{wl1svm}
\alias{wl1svm}
\title{Fit an weighted svm model with given weight pi with L1-norm penalty}
\usage{
wl1svm(
  x.train,
  y.train,
  PI = 0.5,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  lambda = 0.01,
  osqp.option = TRUE,
  beta_precision
)
}
\arguments{
\item{x.train}{Training data in matrix form}

\item{y.train}{Label as vector}

\item{PI}{Weight on negative class data}

\item{kernel}{Kernel to choose, refer to function wsvm.kernel, default is linear kernel}

\item{lambda}{Regulatory parameter for L1-norm penalty and for solving weighted SVMs, lambda > 0}

\item{osqp.option}{choose to use lpSolveAPI or OSQP as solver}

\item{beta_precision}{round decimal points for beta}
}
\value{
L1-norm wSVM model object
}
\description{
Fit an weighted svm model with given weight pi with L1-norm penalty
}
