% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{wsvm.predict}
\alias{wsvm.predict}
\title{predict the label with given wsvm classifier}
\usage{
wsvm.predict(
  x,
  x.train,
  c,
  d,
  kernel = list(type = "linear", param1 = NULL, param2 = NULL)
)
}
\arguments{
\item{x}{New data point}

\item{x.train}{Training data matrix}

\item{c}{Fitted classifier c}

\item{d}{Fitted classifier d}

\item{kernel}{Kernel to use, default 'linear'}
}
\value{
return Predicted label {+1,-1}
}
\description{
predict the label with given wsvm classifier
}
