% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.r
\name{wsvm.prob}
\alias{wsvm.prob}
\title{Probability estimation of given data X}
\usage{
wsvm.prob(
  x.train,
  y.train,
  x.test,
  y.test,
  svm.type = "wl2svm",
  kernel = list(type = "linear", param1 = NULL, param2 = NULL),
  plot.option = "",
  num.plot.imp.features = 10,
  f_threshold = 3,
  lambda,
  lambda2,
  osqp.option,
  beta_precision = 1e-05,
  eps = 1e-10
)
}
\arguments{
\item{x.train}{training data}

\item{y.train}{training label}

\item{x.test}{testing data}

\item{y.test}{testing label}

\item{svm.type}{types of weighted SVMs learning algorithms.
'wl2svm': L2-norm weighted SVMs, without feature selection
'wl1svm': Feature selection first by wL1SVM then wL2SVM (Dual Problem, same lambda for both)
'wl1dsvm': Feature selection first by wL1SVM then wL2SVM (Dual Problem, different lambda for L1 and L2)
'wl1pl2.svm': Feature selection first by wL1SVM then wL2SVM (Primal Problem)
'wl1p.svm': Directly solve beta by wL1SVM (Primal Problem)
'wENsvm': Feature selection and probability estimation directly by elastic net (Dual)
'wENp.svm': Feature selection and probability estimation directly by elastic net (Primal)
'wENl2svm': Feature selection by elastic net follow by L2SVM for probability estimation (Dual)
'wENpl2.svm': Feature selection by elastic net follow by L2SVM for probability estimation (Primal)}

\item{kernel}{Kernel params list. default: 'linear' in sparse learning}

\item{plot.option}{plot feature importance map based on threshold: 'f_threshold' or top N features 'f_importance'}

\item{num.plot.imp.features}{number of important features in plot}

\item{f_threshold}{feature importance threshold value}

\item{lambda}{lambda for L1-norm penalty}

\item{lambda2}{lambda for L2 penalty}

\item{osqp.option}{option to use osqp as solver}

\item{beta_precision}{minimum value of beta before set to 0}

\item{eps}{Small threshold}
}
\value{
Result matrix including estimated probability matrix and feature importance matrix
}
\description{
Probability estimation of given data X
}
